front

+++ backend/src/controllers/dashboard/colorControllers.js
@@ -1,28 +1,27 @@
-import { media } from "../configs/media.js";
-import { Collections } from "../models/index.js";
+import { media } from "../../configs/media.js";
+import { Colors } from "../../models/index.js";
 import fs from "fs";
-
 const index = async (req, res, next) => {
   try {
-    const data = await Collections.find();
+    const data = await Colors.find();
 
-    data.map((item) => (item.banner = media("Collections", item.banner)));
+    data.map((item) => (item.image = media("Colors", item.image)));
 
     return res.status(200).json({ success: true, data: data });
   } catch (error) {
-    next(error);
+    return next(error);
   }
 };
-
 const store = async (req, res, next) => {
+  console.log(req.body);
   try {
     if (!req.file) {
-      throw new Error("Banner collection is not valid!");
+      throw new Error("Image color is not valid!");
     }
 
-    const data = new Collections({
+    const data = new Colors({
       name: req.body.name,
-      banner: req.file.filename,
+      image: req.file.filename,
     });
 
     await data.save();
@@ -30,11 +29,11 @@ const store = async (req, res, next) => {
     return res.status(200).json({
       success: true,
       data: data,
-      message: "Store collection successfully!",
+      message: "Store color successfully!",
     });
   } catch (error) {
     if (req.file) {
-      const filePath = `public/Collections/${req.file.filename}`;
+      const filePath = `public/Colors/${req.file.filename}`;
       fs.existsSync(filePath) && fs.unlinkSync(filePath);
     }
 
@@ -44,11 +43,11 @@ const store = async (req, res, next) => {
 
 const show = async (req, res, next) => {
   try {
-    const data = await Collections.findOne({ slug: req.params.slug });
-    if (!data) {
+    const data = await Colors.findById(req.params.id);
+    if (data) {
       throw new Error("Not found");
     }
-    data.banner = media("Collections", data.banner);
+    data.image = media("Colors", data.image);
     return res.status(200).json({ success: true, data: data });
   } catch (error) {
     return next(error);
@@ -57,18 +56,18 @@ const show = async (req, res, next) => {
 
 const update = async (req, res, next) => {
   try {
-    const data = await Collections.findById(req.body.id);
+    const data = await Colors.findById(req.params.id);
     data.name = req.body.name;
     data.updatedAt = Date.now();
     if (req.file) {
-      const filePath = `public/Collections/${data.banner}`;
+      data.image = req.file.filname;
+      const filePath = `public/Colors/${data.image}`;
       fs.existsSync(filePath) && fs.unlinkSync(filePath);
-      data.banner = req.file.filename;
     }
     await data.save();
     return res
       .status(200)
-      .json({ success: true, message: "Update Collection Successfully!" });
+      .json({ success: true, message: "Update color successfully!" });
   } catch (error) {
     return next(error);
   }
@@ -76,16 +75,15 @@ const update = async (req, res, next) => {
 
 const destroy = async (req, res, next) => {
   try {
-    const data = await Collections.findById(req.params.id);
-
-    const filePath = `public/Collections/${data.banner}`;
-    fs.existsSync(filePath) && fs.unlinkSync(filePath);
-
+    const data = await Colors.findById(req.params.id);
+    if (data.image) {
+      const filePath = `public/Colors/${data.image}`;
+      fs.existsSync(filePath) && fs.unlinkSync(filePath);
+    }
     await data.deleteOne();
-
     return res
       .status(200)
-      .json({ success: true, message: "Delete Collection Successfully!" });
+      .json({ success: true, message: "Delete color successfully!" });
   } catch (error) {
     return next(error);
   }
